requirements.txt
pyserial>=3.4
requests>=2.20.0
opencv-python>=4.5.0
numpy>=1.19.0
tqdm>=4.36.0
matplotlib>=3.4.0
scipy>=1.5.0
Pillow>=8.0.0